Minor Tent Wall Tweaks

+++ src/main/java/rtk/block/BlockTentBreakable.java
@@ -7,12 +7,9 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.block.statemap.StateMap;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemBlock;
-import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.EnumFacing;
-import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import net.minecraftforge.client.model.ModelLoader;
 import rtk.RTK;
 
 public class BlockTentBreakable extends BlockTent {
@@ -64,10 +61,19 @@ public class BlockTentBreakable extends BlockTent {
     public void onBlockDestroyedByPlayer(World world, BlockPos pos, IBlockState state) {
         super.onBlockDestroyedByPlayer(world, pos, state);
 
+//        for(BlockPos neighbor : new BlockPos[]{pos.up(), pos.down(), pos.north(), pos.south(), pos.east(), pos.west()})
+//            BlockTentWall.tryPop(world, null, neighbor);
+
         if(!state.getValue(DEPLOYED))
             return;
 
-        for(BlockPos otherPos : tentCuboid(pos))
-            world.destroyBlock(otherPos, true);
+        for(BlockPos otherPos : tentCuboid(pos)){
+            IBlockState bs = world.getBlockState(otherPos);
+            if(bs.getBlock() instanceof BlockTentWall)
+                BlockTentWall.tryPop(world, otherPos);
+            else
+                world.destroyBlock(otherPos, true);
+        }
+
     }
 }
